ADD general.py
